+++ loguru/_recattrs.py
@@ -1,3 +1,4 @@
+import pickle
 from collections import namedtuple
 
 
@@ -63,5 +64,18 @@ class RecordException(namedtuple("RecordException", ("type", "value", "traceback
         return "(type=%r, value=%r, traceback=%r)" % (self.type, self.value, self.traceback)
 
     def __reduce__(self):
-        exception = (self.type, self.value, None)  # tracebacks are not pickable
-        return (RecordException, exception)
+        try:
+            pickled_value = pickle.dumps(self.value)
+        except pickle.PickleError:
+            return (RecordException, (self.type, None, None))
+        else:
+            return (RecordException._from_pickled_value, (self.type, pickled_value, None))
+
+    @classmethod
+    def _from_pickled_value(cls, type_, pickled_value, traceback_):
+        try:
+            value = pickle.loads(pickled_value)
+        except pickle.PickleError:
+            return cls(type_, None, traceback_)
+        else:
+            return cls(type_, value, traceback_)