+++ tests/test_add_option_enqueue.py
@@ -3,11 +3,12 @@ import pytest
 import time
 import re
 import sys
+import pickle
 
 
 class NotPicklable:
     def __getstate__(self):
-        raise RuntimeError("You shall not serialize me!")
+        raise pickle.PicklingError("You shall not serialize me!")
 
     def __setstate__(self, state):
         pass
@@ -18,7 +19,7 @@ class NotUnpicklable:
         return "..."
 
     def __setstate__(self, state):
-        raise RuntimeError("You shall not de-serialize me!")
+        raise pickle.UnpicklingError("You shall not de-serialize me!")
 
 
 class NotWritable:
@@ -80,7 +81,7 @@ def test_caught_exception_queue_put(writer, capsys):
     assert out == ""
     assert lines[0] == "--- Logging error in Loguru Handler #0 ---"
     assert re.match(r"Record was: \{.*Bye bye.*\}", lines[1])
-    assert lines[-2] == "RuntimeError: You shall not serialize me!"
+    assert lines[-2].endswith("PicklingError: You shall not serialize me!")
     assert lines[-1] == "--- End of logging error ---"
 
 
@@ -98,7 +99,7 @@ def test_caught_exception_queue_get(writer, capsys):
     assert out == ""
     assert lines[0] == "--- Logging error in Loguru Handler #0 ---"
     assert lines[1] == "Record was: None"
-    assert lines[-2] == "RuntimeError: You shall not de-serialize me!"
+    assert lines[-2].endswith("UnpicklingError: You shall not de-serialize me!")
     assert lines[-1] == "--- End of logging error ---"
 
 
@@ -124,7 +125,7 @@ def test_not_caught_exception_queue_put(writer, capsys):
 
     logger.info("It's fine")
 
-    with pytest.raises(RuntimeError, match=r"You shall not serialize me!"):
+    with pytest.raises(pickle.PicklingError, match=r"You shall not serialize me!"):
         logger.bind(broken=NotPicklable()).info("Bye bye...")
 
     logger.remove()
@@ -149,7 +150,7 @@ def test_not_caught_exception_queue_get(writer, capsys):
     assert writer.read() == "It's fine\n"
     assert out == ""
     assert lines[0].startswith("Exception")
-    assert lines[-1] == "RuntimeError: You shall not de-serialize me!"
+    assert lines[-1].endswith("UnpicklingError: You shall not de-serialize me!")
 
 
 def test_not_caught_exception_sink_write(capsys):
@@ -183,3 +184,26 @@ def test_wait_for_all_messages_enqueued(capsys):
 
     assert out == ""
     assert err == "".join("%d\n" % i for i in range(10))
+
+
+@pytest.mark.parametrize("arg", [NotPicklable(), NotUnpicklable()])
+def test_logging_not_picklable_exception(arg):
+    exception = None
+
+    def sink(message):
+        nonlocal exception
+        exception = message.record["exception"]
+
+    logger.add(sink, enqueue=True, catch=False)
+
+    try:
+        raise ValueError(arg)
+    except Exception:
+        logger.exception("Oups")
+
+    logger.remove()
+
+    type_, value, traceback_ = exception
+    assert type_ is ValueError
+    assert value is None
+    assert traceback_ is None